security changes added

+++ src/main/java/com/spring/EventManagementSystem/service/UserServiceLayerImpl.java
@@ -1,7 +1,7 @@
 package com.spring.EventManagementSystem.service;
 
 import com.spring.EventManagementSystem.component.CSVDataLoader;
-import com.spring.EventManagementSystem.dao.UserJPARepository;
+import com.spring.EventManagementSystem.repository.UserJPARepository;
 import com.spring.EventManagementSystem.entity.Users;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,7 +27,7 @@ public class UserServiceLayerImpl implements UserServiceLayer{
     }
 
     @Override
-    public Users findById(int id){
+    public Users findById(Long id){
 
         Optional<Users> userList = myUserJPARepository.findById(id);
         if(userList.isPresent()){
